requirements.txt
SpeechRecognition
pyttsx3
wikipedia
requests
google-generativeai
pyautogui
pyaudio
pytube
python-dotenv
pyaudio